[Build] Use published marks in another place more
The goal is to stop accessing tasks in other projects, so we use dummy files.
Eventually we could use the local folders where the artifacts are published.
